requirements.txt
python-telegram-bot==13.7
paramiko
psycopg2